Fix type of `StatusBuilder::media_ids`
The identifier in the JSON response to a POST to `/media` is a string,
as can be seen in the Attachment entity.

Even though that identifier has always been a string of numbers (at
least in Mastodon's case), it should not be assumed that it will
continue to be so in the future, so keeping it as a String seems more
prudent.

Additionally, this simplifies the process of posting a new status with
an attachment, as it is no longer necessary to parse the resulting
attachment's id to an u64.