Remove anti-pattern Into<> impl
First of all, this is an anti-pattern. See

    https://rust-lang.github.io/rust-clippy/master/index.html#from_over_into

next: What the hell? Into<Option<_>> implemented on a mutable reference
that does not even need to be mutable? I don't see why this would ever be
required.

It seems, from what I see, that this is not even required in the code.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>