Remove traits from equation
The `MastodonClient` and `MastodonUnauthenticated` did not serve a real
purpose in the codebase of the crate.

The idea that they can be reimplemented for testing purposes is
valid, but the mechanic to do that via a trait is overkill.

The refactoring will introduce an replacement for that at some point in
the future.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>